Apache 2.0 License?
You don't have a LICENSE file yet.  But, I could see wanting to make PRs or fork this for something educational I'm working on.  Apache 2.0 is what I'm suggesting as it seems to be an increasingly popular licenses, especially in terms of companies picking the code up for use without getting hassled by internal lawyers.  For example, gpt-oss was recently released under Apache 2.0